requirements.txt
absl-py==1.3.0
alembic==1.10.2
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
click==8.1.3
cmaes==0.9.1
colorlog==6.7.0
contourpy==1.0.6
cycler==0.11.0
ete3==3.1.2
filelock==3.12.0
fonttools==4.38.0
fsspec==2023.5.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.23.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
imutils==0.5.4
joblib==1.2.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
networkx==2.8.8
nltk==3.8.1
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.8.0.74
optuna==3.1.0
packaging==22.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pyparsing==3.0.9
pyrealsense2==2.54.1.5217
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
segment-anything==1.0
sentencepiece==0.1.99
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
SQLAlchemy==2.0.5.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
torchviz==0.0.2
tqdm==4.64.1
transformers==4.29.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
yacs==0.1.8
zipp==3.11.0
